Change elem size type from int to long
git-svn-id: https://www.brockmann-consult.de/svn/os/beam-4.x/trunk@3024 d90d2d48-2b33-0410-a973-e72f85796a02

+++ beam-core/src/main/java/org/esa/beam/framework/datamodel/Band.java
@@ -419,10 +419,10 @@ public class Band extends AbstractBand {
         long size = 0L;
         if (isPartOfSubset(subsetDef)) {
             size += 256; // add estimated overhead of 256 bytes
-            int numDataElems = getNumDataElems();
+            long numDataElems = getNumDataElems();
             if (subsetDef != null) {
-                int width = getSceneRasterWidth();
-                int height = getSceneRasterHeight();
+                long width = getSceneRasterWidth();
+                long height = getSceneRasterHeight();
                 Rectangle region = subsetDef.getRegion();
                 if (region != null) {
                     width = region.width;
